beamer: move create_request_id to beamer.util
This is going to be used outside of tests too.
Also add typing information.
